Move much of the shared code from command_tasks on engine and application to command task base
