+++ railties/lib/rails/commands/commands_tasks.rb
@@ -1,4 +1,5 @@
 require 'rails/commands/rake_proxy'
+require 'rails/commands/commands_tasks_base'
 
 module Rails
   # This is a class which takes in a rails command and initiates the appropriate
@@ -6,10 +7,7 @@ module Rails
   #
   # Warning: This class mutates ARGV because some commands require manipulating
   # it before they are run.
-  class CommandsTasks # :nodoc:
-    include Rails::RakeProxy
-
-    attr_reader :argv
+  class CommandsTasks < CommandsTasksBase# :nodoc:
 
     HELP_MESSAGE = <<-EOT
 Usage: rails COMMAND [ARGS]
@@ -38,30 +36,8 @@ EOT
 
     COMMAND_WHITELIST = %w(plugin generate destroy console server dbconsole runner new version help test)
 
-    def initialize(argv)
-      @argv = argv
-    end
-
-    def run_command!(command)
-      command = parse_command(command)
-
-      if COMMAND_WHITELIST.include?(command)
-        send(command)
-      else
-        run_rake_task(command)
-      end
-    end
-
     def plugin
       require_command!("plugin")
-    end
-
-    def generate
-      generate_or_destroy(:generate)
-    end
-
-    def destroy
-      generate_or_destroy(:destroy)
     end
 
     def console
@@ -91,10 +67,6 @@ EOT
       end
     end
 
-    def test
-      require_command!("test")
-    end
-
     def dbconsole
       require_command!("dbconsole")
       Rails::DBConsole.start
@@ -112,11 +84,6 @@ EOT
       end
     end
 
-    def version
-      argv.unshift '--version'
-      require_command!("application")
-    end
-
     def help
       write_help_message
       write_commands ADDITIONAL_COMMANDS + formatted_rake_tasks
@@ -132,10 +99,6 @@ EOT
 
       def shift_argv!
         argv.shift if argv.first && argv.first[0] != '-'
-      end
-
-      def require_command!(command)
-        require "rails/commands/#{command}"
       end
 
       def generate_or_destroy(command)
@@ -161,20 +124,8 @@ EOT
         puts HELP_MESSAGE
       end
 
-      def write_commands(commands)
-        width = commands.map { |name, _| name.size }.max || 10
-        commands.each { |command| printf(" %-#{width}s   %s\n", *command) }
-      end
-
-      def parse_command(command)
-        case command
-        when '--version', '-v'
-          'version'
-        when '--help', '-h'
-          'help'
-        else
-          command
-        end
+      def command_whitelist
+        COMMAND_WHITELIST
       end
   end
 end
