+++ railties/lib/rails/engine/commands_tasks.rb
@@ -1,11 +1,9 @@
 require 'rails/commands/rake_proxy'
+require 'rails/commands/commands_tasks_base'
 
 module Rails
   class Engine
-    class CommandsTasks # :nodoc:
-      include Rails::RakeProxy
-
-      attr_reader :argv
+    class CommandsTasks < CommandsTasksBase # :nodoc:
 
       HELP_MESSAGE = <<-EOT
 Usage: rails COMMAND [ARGS]
@@ -26,47 +24,12 @@ In addition to those commands, there are:
 
       COMMAND_WHITELIST = %w(generate destroy version help test)
 
-      def initialize(argv)
-        @argv = argv
-      end
-
-      def run_command!(command)
-        command = parse_command(command)
-
-        if COMMAND_WHITELIST.include?(command)
-          send(command)
-        else
-          run_rake_task(command)
-        end
-      end
-
-      def generate
-        generate_or_destroy(:generate)
-      end
-
-      def destroy
-        generate_or_destroy(:destroy)
-      end
-
-      def test
-        require_command!("test")
-      end
-
-      def version
-        argv.unshift '--version'
-        require_command!("application")
-      end
-
       def help
         write_help_message
         write_commands(formatted_rake_tasks)
       end
 
       private
-
-        def require_command!(command)
-          require "rails/commands/#{command}"
-        end
 
         def generate_or_destroy(command)
           load_generators
@@ -86,22 +49,6 @@ In addition to those commands, there are:
           puts HELP_MESSAGE
         end
 
-        def write_commands(commands)
-          width = commands.map { |name, _| name.size }.max || 10
-          commands.each { |command| printf(" %-#{width}s   %s\n", *command) }
-        end
-
-        def parse_command(command)
-          case command
-          when '--version', '-v'
-            'version'
-          when '--help', '-h'
-            'help'
-          else
-            command
-          end
-        end
-
         def rake_tasks
           return @rake_tasks if defined?(@rake_tasks)
 
@@ -110,6 +57,10 @@ In addition to those commands, there are:
           Rake.application.init('rails')
           Rake.application.load_rakefile
           @rake_tasks = Rake.application.tasks.select(&:comment)
+        end
+
+        def command_whitelist
+          COMMAND_WHITELIST
         end
     end
   end
